Moved LEB functions from utils.cpp to crypto.cpp
